requirements.txt
discord.py-self
python-telegram-bot
